pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waferlab"
version = "0.1.0"
description = "Deep learning lab for wafer map defect analysis"
requires-python = ">=3.10"
dependencies = [
    "kagglehub",
    "numpy",
    "pandas",
    "h5py",
    "pyarrow",
    "PyYAML",
    "matplotlib",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]
